don't attempt incircle test against a ghost point
previously only checked if an edge contained ghost points when checking for legality. it could not but the other point might be, which is the point that is tested for location inside a triangle's circumcircle, which doesn't make sense for a ghost point. they are always supposed to lie beyond all circumcircles, so add a check and exit early to say if a test point is a ghost point, the incircle "result" is that it always passes the test.